+++ admin/plugin_log.php
@@ -3,7 +3,7 @@ namespace TLC\TTSurvey;
 
 if( !current_user_can('manage_options') ) { wp_die('Unauthorized user'); }
 
-require_once plugin_path('logger.php');
+require_once plugin_path('include/logger.php');
 require_once plugin_path('settings.php');
 
 $nonce = wp_nonce_field(OPTIONS_NONCE);
